+++ src/entities/UploadImageEntities.ts
@@ -1,0 +1,28 @@
+export interface UploadImageEntity {
+  resourceId: string;
+  resource: string;
+  category: string;
+  file: any;
+  path: string;
+}
+
+export interface ImageEntity extends UploadImageEntity {
+  id: string;
+}
+
+export const ImageEntity_INTI: ImageEntity = {
+  id: "",
+  resourceId: "",
+  resource: "",
+  category: "",
+  file: "",
+  path: "",
+};
+
+export const UploadImageEntity_INTI: UploadImageEntity = {
+  resourceId: "",
+  resource: "",
+  category: "",
+  file: "",
+  path: "",
+};